fixed progress bars in arenas

+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/blocks/EssenceBlockRed.java
@@ -155,7 +155,7 @@ public class EssenceBlockRed extends EssenceBlock {
             }
         }
 
-        float newProgress = entitiesKilled / ENTITIES_TO_KILL;
+        float newProgress = (float) entitiesKilled / ENTITIES_TO_KILL;
         essenceBlockEntity.getEventBar().setProgress(1 - newProgress);
         essenceBlockEntity.setExtraEventTrackingProgress(entitiesKilled);
         if (entitiesKilled == ENTITIES_TO_KILL) {
